Roles/tailscale server (#1)
* fix: Made yaml valid ansible file

File was missing the 3 hyphens marking it as an ansible file. updated file to correctly have that required line.

* feat: Updated inventory to include new tailscale groups

added 3 new tailscale related groups to inventory file. This allows for more granular control of tailscale settings.

* feat: Add install tasks for tailscale for common

Added install tasks for both linux and macos for tailscale.
tailscale setup will happen per role in a later feature branch

* feat: Added all.yaml file

added new all.yaml group_var file. This includes a variable containing the tailscale login url used by all devices. This is encrypted with ansible vault.

* feat: added new role tailscale_server

This new role is for servers that need tailscale and connect with a preauth key, but do not function as a subnet router.
This role simply starts tailscale, with the required preauth key defined in group_vars

* update: updated site.yaml to include new role

update site.yaml to include the new tailscale_server role.