feat: add new subnet router role
Added new role tailscale_subnet_router. This role is for devices that will be using tailscale to advertise subnet routes to the rest of the tailnet.
Includes new task to run tailscale up with required flags.
also includes var file containing the current subnet to advertise. currently this applies to all subnet router as only one device uses this role.